requirements.txt
Django==4.2
djangorestframework==3.14.0
django-environ==0.11.2